Create module Lazy in Util Remove validator registration and leverage Zeitwerk
