+++ lib/grape/http/headers.rb
@@ -34,7 +34,7 @@ module Grape
       OPTIONS = 'OPTIONS'
 
       SUPPORTED_METHODS = [GET, POST, PUT, PATCH, DELETE, HEAD, OPTIONS].freeze
-      SUPPORTED_METHODS_WITHOUT_OPTIONS = Grape::Util::LazyObject.new { [GET, POST, PUT, PATCH, DELETE, HEAD].freeze }
+      SUPPORTED_METHODS_WITHOUT_OPTIONS = Grape::Util::Lazy::Object.new { [GET, POST, PUT, PATCH, DELETE, HEAD].freeze }
 
       HTTP_ACCEPT_VERSION    = 'HTTP_ACCEPT_VERSION'
       HTTP_TRANSFER_ENCODING = 'HTTP_TRANSFER_ENCODING'
@@ -42,7 +42,7 @@ module Grape
 
       FORMAT                 = 'format'
 
-      HTTP_HEADERS = Grape::Util::LazyObject.new do
+      HTTP_HEADERS = Grape::Util::Lazy::Object.new do
         common_http_headers = %w[
           Version
           Host
