--- lib/grape/util/lazy_value_array.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-module Grape
-  module Util
-    class LazyValueArray < LazyValueEnumerable
-      def initialize(array)
-        super
-        @value_hash = []
-        array.each_with_index do |value, index|
-          self[index] = value
-        end
-      end
-
-      def evaluate
-        @value_hash.map(&:evaluate)
-      end
-    end
-  end
-end
