--- lib/grape/util/lazy_value_hash.rb
@@ -1,20 +1,0 @@
-# frozen_string_literal: true
-
-module Grape
-  module Util
-    class LazyValueHash < LazyValueEnumerable
-      def initialize(hash)
-        super
-        @value_hash = ActiveSupport::HashWithIndifferentAccess.new
-        hash.each do |key, value|
-          self[key] = value
-        end
-      end
-
-      def evaluate
-        @value_hash.transform_values(&:evaluate)
-      end
-    end
-
-  end
-end
